# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "News Aggregator",
        "description": "API documentation for news aggregator",
        "contact": {
            "email": "deekshithshetty48@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/article/list": {
            "get": {
                "tags": [
                    "Article"
                ],
                "summary": "Fetch a list of articles",
                "operationId": "d92299097b35e00ae4bc0f3beae4e99a",
                "parameters": [
                    {
                        "name": "search_key",
                        "in": "query",
                        "description": "Key to search articles by title or content.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "category",
                        "in": "query",
                        "description": "Filter articles by category.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "source",
                        "in": "query",
                        "description": "Filter articles by source name.",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "published_date",
                        "in": "query",
                        "description": "Filter articles by published date.",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "date"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number for pagination.",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful response with article list",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "article list fetch successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "data": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 11
                                                            },
                                                            "source_name": {
                                                                "type": "string",
                                                                "example": "BBC News"
                                                            },
                                                            "author": {
                                                                "type": "string",
                                                                "example": "BBC News"
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Budget 2024: Minimum wage to rise to £12.21 an hour next year"
                                                            },
                                                            "published_at": {
                                                                "type": "string",
                                                                "format": "date-time",
                                                                "example": "2024-10-29T18:37:17Z"
                                                            },
                                                            "category": {
                                                                "type": "string",
                                                                "example": "none"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "first_page_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8083/api/article/list?page=1"
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8083/api/article/list?page=1"
                                                },
                                                "links": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "url": {
                                                                "type": "string",
                                                                "nullable": true
                                                            },
                                                            "label": {
                                                                "type": "string"
                                                            },
                                                            "active": {
                                                                "type": "boolean"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "path": {
                                                    "type": "string",
                                                    "example": "http://localhost:8083/api/article/list"
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "prev_page_url": {
                                                    "type": "string",
                                                    "nullable": true
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 2
                                                },
                                                "total": {
                                                    "type": "integer",
                                                    "example": 2
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "500": {
                        "description": "Internal server error"
                    }
                },
                "security": [
                    {
                        "bearer": [
                            "Token"
                        ]
                    }
                ]
            }
        },
...
    },
...
}